changed the data type from integr to long for the jpa repository

+++ src/main/java/com/foodapi/betaecommerceapiv2/repository/order/OrderItemsRepository.java
@@ -3,5 +3,5 @@ package com.foodapi.betaecommerceapiv2.repository.order;
 import com.foodapi.betaecommerceapiv2.models.order.OrderItem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OrderItemsRepository extends JpaRepository<OrderItem, Integer> {
+public interface OrderItemsRepository extends JpaRepository<OrderItem, Long> {
 }
